# Use plain function pointers for type info accessors.