It turns out that std::function not only adds runtime overhead, but also
that it is significantly slower to compiler than a function pointer.

This commit partially addresses #546.