@bbannier 2020-10-22
> Are you seeing a difference in compile times?

I tested with by compiling `hilti/runtime/src/tests/type-info.cc` which includes moderate use of this header and see a significant, but only moderate speedup:

- before: 3.617 s ±  0.018 s
- after: 3.505 s ±  0.033 s

Looking at the time profile a bigger win be possible from getting rid of the `std::variant` use in this header. I'll look into that next.