@bbannier 2020-10-22
> Try comparing it for one of the btests, say `spicy/types/unit/sub-unit.spicy`

For that one the difference is bigger:

- before: (11.979 s ± 0.610) s
- after: (9.793 s ± 0.242) s

FTR, I first safed the output to a file, and then benchmarked generating an object file from that source file
```.sh
spicyc -c spicy/types/unit/sub-unit.spicy -o foo.cc
$(spicy-config --cxx --cxxflags --ldflags) -c foo.cc -o foo.o
```